Fix warning and Win32 in general

+++ lib/sj4common/bsd-snprintf.c
@@ -364,7 +364,7 @@ static size_t dopr(char* buffer, size_t maxlen, const char* format, va_list args
 				break;
 			case 'p':
 				strvalue = va_arg(args, void*);
-				fmtint(buffer, &currlen, maxlen, (long)strvalue, 16, min, max, flags);
+				fmtint(buffer, &currlen, maxlen, (size_t)strvalue, 16, min, max, flags);
 				break;
 			case 'n':
 				if(cflags == DP_C_SHORT) {
